fixed crash

+++ src/main/java/net/rbm/cybercraft/procedures/AdrenalineBoosterEffectProcedure.java
@@ -45,7 +45,7 @@ public class AdrenalineBoosterEffectProcedure {
 							|| (sourceentity instanceof LivingEntity _livEnt ? _livEnt.getMainHandItem() : ItemStack.EMPTY).getItem() instanceof PickaxeItem
 							|| (sourceentity instanceof LivingEntity _livEnt ? _livEnt.getMainHandItem() : ItemStack.EMPTY).getItem() instanceof ShovelItem)) {
 				if (sourceentity instanceof Player _player)
-					_player.getFoodData().setFoodLevel((int) ((sourceentity instanceof Player _plr ? _plr.getFoodData().getFoodLevel() : 0) + 1));
+					_player.getFoodData().setFoodLevel((sourceentity instanceof Player _plr ? _plr.getFoodData().getFoodLevel() : 0) + 1);
 			}
 		}
 	}
